Update signals logging + proper throttle controlling logic

+++ signalHandler/outputSignals.c
@@ -3,8 +3,8 @@
 // All output signal definitions here
 
 /* --- Motor throttle signals, same value for both motors */
-uint16_t outputSignal_motor1_throttle;
-uint16_t outputSignal_motor2_throttle;
+uint8_t outputSignal_motor1_throttle;
+uint8_t outputSignal_motor2_throttle;
 
 /* --- Software level battery protection signals --- */
 bool outputSignal_battery1_currentTooHigh;
